\name{CRD1}
\alias{CRD1}

\title{
  Completely Randomized Design (CRD)
}

\description{
  A \code{list} illustrating the resources of \code{TukeyC} package
  related to Completely Randomized Design (\samp{CRD}).
}

\usage{
  data(CRD1)
}

\details{
  A simulated data to model a Completely Randomized Design (\samp{CRD}) of 4 
  factor levels and 6 repetitions.
}

\keyword{datasets}
