\name{SPE}
\alias{SPE}

\title{
  Split-plot Experiment (SPE)
}

\description{
  A \code{list} to illustrate the resources of \code{TukeyC} package
  related to Split-plot Experiment (\samp{SPE}).
}

\usage{
  data(SPE)
}

\details{
  A simulated data to model a Split-plot Experiment (\samp{SPE}) with 3 plots, each
  one split 4 times and 6 repetitions per split.
}

\keyword{datasets}
