\name{SPET}
\alias{SPET}

\title{
  Split-plot Experiment in Time (SPET)}

\description{
  The experiment consists of 8 treatments (7 leguminous cover crops and maize) in a 
  Randomized Complete Block Design (\samp{RCBD}) and the yield by plot (kg/plot).}

\usage{
  data(SPET)
}

\source{
  Gomes, F.P. (1990).
  Curso de Estatistica Experimental. 13 ed. Editora NOBEL, Piracicaba,
  Brazil, page 157.}

\keyword{datasets}
