\name{SSPE}
\alias{SSPE}

\title{
  Split-split-plot Experiment (SSPE)
}

\description{
  A \code{list} to illustrate the resources of \code{TukeyC} package
  related to Split-split-plot Experiment (\samp{SSPE}).
}

\usage{
  data(SSPE)
}

\details{
  A simulated data to model a Split-split-plot Experiment (\samp{SSPE}) with 3 plots,
  each one split 3 times, each split, split again 5 times and 
  4 repetitions per split-split.
}

\keyword{datasets}
