\name{xtable}
\alias{xtable}

\title{Create a Table for Export}

\description{
  This function is re-exported from the \pkg{xtable} package so that
  \code{xtable()} is available after \code{library(TukeyC)} without
  requiring a separate \code{library(xtable)} call.

  For \code{TukeyC} objects the S3 method \code{\link{xtable.TukeyC}} is dispatched
  automatically.  For full documentation of the generic see
  \code{help("xtable", package = "xtable")}.
}

\seealso{
  \code{\link{xtable.TukeyC}}, \code{\link[xtable]{xtable}}
}

\keyword{utilities}
