\name{CRD2}
\alias{CRD2}

\title{Completely Randomized Design (CRD)}

\description{
  A \code{list}illustrating the resources of \code{TukeyC} package
  related to Completely Randomized Design (CRD).}

\usage{
  data(CRD2)}

\details{
  A simulated data to model a Completely Randomized Design (CRD) of 45 
  factor levels and 4 repetitions.}

\keyword{datasets}
