\name{FE}
\alias{FE}

\title{Factorial Experiment (FE)}

\description{
  A \code{list} illustrating  the resources of \code{TukeyC} package
  related to Factorial Experiment (FE).}

\usage{
  FE}

\details{
  A simulated data to model a Factorial Experiment (FE) with 3 factors, 
  2 levels per factor and 4 blocks.}

\keyword{datasets}
