\name{RCBD}
\alias{RCBD}

\title{Randomized Complete Block Design (RCBD)}

\description{
  A \code{list} illustrating  the resources of \code{TukeyC} package
  related to Randomized Complete Block Design (\samp{RCBD}).}

\usage{
  data(RCBD)}

\details{
  A simulated data to model a Randomized Complete Block Design (\samp{RCBD}) of 5 
  factor levels, 4 blocks and 4 factor levels repetitions one in each block.}

\keyword{datasets}
